refactor(static-image): extract presentation cache and preload modules
