+++ src/app/overlays/images/mod.rs
@@ -1,17 +1,16 @@
+mod cache;
+mod preload;
+mod present;
 mod state;
 mod types;
 
 use super::super::*;
-use super::inline_image::{
-    ImageProtocol, OverlayPresentState, RenderedImageDimensions, TerminalWindowSize,
-    fit_image_area, place_terminal_image, preview_log, read_png_dimensions,
-};
-use anyhow::Result;
+use super::inline_image::{RenderedImageDimensions, TerminalWindowSize, read_png_dimensions};
 use image::{DynamicImage, GenericImageView, ImageFormat, ImageReader, imageops::FilterType};
 use quick_xml::{Reader, events::Event};
 use ratatui::layout::Rect;
 use std::{
-    collections::{HashSet, hash_map::DefaultHasher},
+    collections::hash_map::DefaultHasher,
     env, fs,
     fs::File,
     hash::{Hash, Hasher},
@@ -22,7 +21,6 @@ use std::{
     time::SystemTime,
 };
 
-use self::types::{DisplayedStaticImagePreview, StaticImagePreloadViewport};
 pub(in crate::app) use self::types::{
     ImagePreviewState, PreparedStaticImage, PreparedStaticImageAsset, StaticImageKey,
     StaticImageOverlayMode, StaticImageOverlayPreparation, StaticImageOverlayRequest,
@@ -39,178 +37,6 @@ const FAST_FORCE_RENDER_FFMPEG_RASTER_ARGS: [&str; 4] =
 const DEFAULT_FFMPEG_RASTER_ARGS: [&str; 0] = [];
 
 impl App {
-    pub(in crate::app) fn present_static_image_overlay(
-        &mut self,
-        protocol: ImageProtocol,
-        excluded: &[Rect],
-        out: &mut Vec<u8>,
-    ) -> Result<OverlayPresentState> {
-        let Some(request) = self.active_static_image_overlay_request() else {
-            preview_log("present_static_image_overlay: no request");
-            return Ok(OverlayPresentState::NotRequested);
-        };
-        preview_log(format_args!(
-            "present_static_image_overlay: path={:?} area={:?}",
-            request.path, request.area
-        ));
-        if !self.image_selection_activation_ready() {
-            preview_log("present_static_image_overlay: activation not ready → Waiting");
-            return Ok(OverlayPresentState::Waiting);
-        }
-
-        let prepared = match self.prepared_static_image_for_overlay(&request) {
-            StaticImageOverlayPreparation::Ready(prepared) => prepared,
-            StaticImageOverlayPreparation::Pending => {
-                preview_log("present_static_image_overlay: preparation Pending → Waiting");
-                return Ok(OverlayPresentState::Waiting);
-            }
-            StaticImageOverlayPreparation::Failed => {
-                preview_log("present_static_image_overlay: preparation Failed");
-                self.mark_static_image_failed(&request);
-                self.refresh_preview();
-                return Ok(OverlayPresentState::NotRequested);
-            }
-        };
-        let Some(window_size) = self.cached_terminal_window() else {
-            preview_log("present_static_image_overlay: no cached window size → failed");
-            self.mark_static_image_failed(&request);
-            self.refresh_preview();
-            return Ok(OverlayPresentState::NotRequested);
-        };
-        let placement = self.static_image_display_area(&request, prepared.dimensions, window_size);
-        preview_log(format_args!(
-            "present_static_image_overlay: dims={}x{} placement={:?}",
-            prepared.dimensions.width_px, prepared.dimensions.height_px, placement
-        ));
-        let displayed = DisplayedStaticImagePreview::from_request(
-            &request,
-            placement,
-            self.static_image_clear_area(&request),
-        );
-        let image_changed = self.image_preview.displayed.as_ref() != Some(&displayed);
-        let excluded_changed = excluded != self.image_preview.displayed_excluded.as_slice();
-        if !image_changed && !excluded_changed {
-            preview_log("present_static_image_overlay: already displayed → Displayed");
-            return Ok(OverlayPresentState::Displayed);
-        }
-        // Only clear the old image when the image itself changed, not when only the
-        // excluded rects changed, to avoid a visible flash as the image disappears
-        // and then immediately reappears.
-        if image_changed {
-            out.extend(self.clear_preview_overlay()?);
-        }
-        match place_terminal_image(
-            protocol,
-            &prepared.display_path,
-            placement,
-            excluded,
-            prepared.inline_payload.as_deref(),
-        ) {
-            Ok(bytes) => {
-                preview_log(format_args!(
-                    "present_static_image_overlay: placed {} bytes via {protocol:?}",
-                    bytes.len()
-                ));
-                out.extend(bytes);
-            }
-            Err(e) => {
-                preview_log(format_args!(
-                    "present_static_image_overlay: place_terminal_image error: {e}"
-                ));
-                self.mark_static_image_failed(&request);
-                self.refresh_preview();
-                return Ok(OverlayPresentState::NotRequested);
-            }
-        }
-
-        self.image_preview.displayed = Some(displayed);
-        self.image_preview.displayed_excluded = excluded.to_vec();
-        Ok(OverlayPresentState::Displayed)
-    }
-
-    pub(in crate::app) fn present_preview_visual_overlay(
-        &mut self,
-        protocol: ImageProtocol,
-        excluded: &[Rect],
-        out: &mut Vec<u8>,
-    ) -> Result<OverlayPresentState> {
-        let Some(request) = self.active_preview_visual_overlay_request() else {
-            preview_log("present_preview_visual_overlay: no request");
-            return Ok(OverlayPresentState::NotRequested);
-        };
-        preview_log(format_args!(
-            "present_preview_visual_overlay: path={:?} area={:?}",
-            request.path, request.area
-        ));
-        if !self.image_selection_activation_ready() {
-            preview_log("present_preview_visual_overlay: activation not ready → Waiting");
-            return Ok(OverlayPresentState::Waiting);
-        }
-
-        let prepared = match self.prepared_static_image_for_overlay(&request) {
-            StaticImageOverlayPreparation::Ready(prepared) => prepared,
-            StaticImageOverlayPreparation::Pending => {
-                preview_log("present_preview_visual_overlay: preparation Pending → Waiting");
-                return Ok(OverlayPresentState::Waiting);
-            }
-            StaticImageOverlayPreparation::Failed => {
-                preview_log("present_preview_visual_overlay: preparation Failed");
-                self.mark_static_image_failed(&request);
-                return Ok(OverlayPresentState::NotRequested);
-            }
-        };
-        let Some(window_size) = self.cached_terminal_window() else {
-            preview_log("present_preview_visual_overlay: no cached window size → failed");
-            self.mark_static_image_failed(&request);
-            return Ok(OverlayPresentState::NotRequested);
-        };
-        let placement = self.static_image_display_area(&request, prepared.dimensions, window_size);
-        preview_log(format_args!(
-            "present_preview_visual_overlay: dims={}x{} placement={:?}",
-            prepared.dimensions.width_px, prepared.dimensions.height_px, placement
-        ));
-        let displayed = DisplayedStaticImagePreview::from_request(
-            &request,
-            placement,
-            self.static_image_clear_area(&request),
-        );
-        let image_changed = self.image_preview.displayed.as_ref() != Some(&displayed);
-        let excluded_changed = excluded != self.image_preview.displayed_excluded.as_slice();
-        if !image_changed && !excluded_changed {
-            preview_log("present_preview_visual_overlay: already displayed → Displayed");
-            return Ok(OverlayPresentState::Displayed);
-        }
-        if image_changed {
-            out.extend(self.clear_preview_overlay()?);
-        }
-        match place_terminal_image(
-            protocol,
-            &prepared.display_path,
-            placement,
-            excluded,
-            prepared.inline_payload.as_deref(),
-        ) {
-            Ok(bytes) => {
-                preview_log(format_args!(
-                    "present_preview_visual_overlay: placed {} bytes via {protocol:?}",
-                    bytes.len()
-                ));
-                out.extend(bytes);
-            }
-            Err(e) => {
-                preview_log(format_args!(
-                    "present_preview_visual_overlay: place_terminal_image error: {e}"
-                ));
-                self.mark_static_image_failed(&request);
-                return Ok(OverlayPresentState::NotRequested);
-            }
-        }
-
-        self.image_preview.displayed = Some(displayed);
-        self.image_preview.displayed_excluded = excluded.to_vec();
-        Ok(OverlayPresentState::Displayed)
-    }
-
     pub(in crate::app) fn prepared_static_image_for_overlay(
         &mut self,
         request: &StaticImageOverlayRequest,
@@ -256,345 +82,6 @@ impl App {
             dimensions,
             inline_payload: None,
         })
-    }
-
-    pub(in crate::app) fn refresh_static_image_preloads(&mut self) {
-        let current_static = self.active_static_image_overlay_request();
-        let current_preview_visual = self.active_preview_visual_overlay_request();
-        let current = current_static
-            .as_ref()
-            .cloned()
-            .or(current_preview_visual.as_ref().cloned());
-        let nearby = if let Some(request) = current_static.as_ref() {
-            self.nearby_static_image_overlay_requests(Some(request))
-        } else if current_preview_visual.is_some() {
-            let mut requests = self.nearby_comic_preview_visual_overlay_requests();
-            requests.extend(self.nearby_epub_preview_visual_overlay_requests());
-            requests
-        } else {
-            Vec::new()
-        };
-        let desired = current
-            .iter()
-            .map(StaticImageKey::from_request)
-            .chain(nearby.iter().map(StaticImageKey::from_request))
-            .collect::<HashSet<_>>();
-        self.image_preview
-            .pending_prepares
-            .retain(|key| desired.contains(key));
-        let current_job = current
-            .as_ref()
-            .map(|request| self.image_prepare_request_for_overlay(request));
-        let nearby_jobs = nearby
-            .iter()
-            .map(|request| self.image_prepare_request_for_overlay(request))
-            .collect::<Vec<_>>();
-        self.scheduler
-            .retain_image_prepares(current_job.as_ref(), &nearby_jobs);
-
-        if let Some(request) = current.as_ref()
-            && self.static_image_requires_prepare(request)
-        {
-            self.ensure_static_image_preload(request, jobs::ImageJobPriority::Current);
-        }
-        for request in &nearby {
-            if self.static_image_requires_prepare(request) {
-                self.ensure_static_image_preload(request, jobs::ImageJobPriority::Nearby);
-            }
-        }
-    }
-
-    pub(in crate::app) fn refresh_static_image_preloads_if_needed(&mut self) {
-        let viewport = StaticImagePreloadViewport {
-            selected: self.selected,
-            scroll_row: self.scroll_row,
-            cols: self.frame_state.metrics.cols.max(1),
-            rows_visible: self.frame_state.metrics.rows_visible.max(1),
-            preview_content_area: self.frame_state.preview_content_area,
-            preview_media_area: self.frame_state.preview_media_area,
-            protocol: self.terminal_images.protocol,
-            window: self.cached_terminal_window(),
-        };
-        if self.image_preview.preload_viewport == Some(viewport) {
-            return;
-        }
-        self.image_preview.preload_viewport = Some(viewport);
-        self.refresh_static_image_preloads();
-    }
-
-    pub(in crate::app) fn apply_image_prepare_build(
-        &mut self,
-        build: jobs::ImagePrepareBuild,
-    ) -> bool {
-        let key = StaticImageKey::from_parts(
-            build.path.clone(),
-            build.size,
-            build.modified,
-            build.target_width_px,
-            build.target_height_px,
-            build.force_render_to_cache,
-            build.prepare_inline_payload,
-        );
-        self.image_preview.pending_prepares.remove(&key);
-        let is_current = self
-            .active_static_image_overlay_request()
-            .as_ref()
-            .is_some_and(|request| StaticImageKey::from_request(request) == key)
-            || self
-                .active_preview_visual_overlay_request_unchecked()
-                .as_ref()
-                .is_some_and(|request| StaticImageKey::from_request(request) == key);
-        if build.canceled {
-            self.refresh_static_image_preloads();
-            return is_current;
-        }
-
-        match build.result {
-            Some(prepared) => {
-                self.image_preview.failed_images.remove(&key);
-                self.image_preview
-                    .dimensions
-                    .insert(key.clone(), prepared.dimensions);
-                if let Some(payload) = prepared.inline_payload {
-                    self.remember_static_image_inline_payload(key.clone(), payload);
-                }
-                if prepared.display_path != build.path {
-                    self.remember_rendered_static_image(key, prepared.display_path);
-                }
-                self.refresh_static_image_preloads();
-                is_current
-            }
-            None => {
-                self.image_preview.failed_images.insert(key);
-                if is_current {
-                    self.refresh_preview();
-                    true
-                } else {
-                    false
-                }
-            }
-        }
-    }
-
-    fn cached_static_image_display_path(&mut self, key: &StaticImageKey) -> Option<PathBuf> {
-        if let Some(path) = self.image_preview.rendered_images.get(key)
-            && path.exists()
-        {
-            return Some(path.clone());
-        }
-
-        self.image_preview.rendered_images.remove(key);
-        self.image_preview
-            .render_order
-            .retain(|queued| queued != key);
-        None
-    }
-
-    fn cached_static_image_inline_payload(&self, key: &StaticImageKey) -> Option<Arc<str>> {
-        self.image_preview.inline_payloads.get(key).cloned()
-    }
-
-    fn remember_static_image_inline_payload(&mut self, key: StaticImageKey, payload: Arc<str>) {
-        self.image_preview
-            .inline_payloads
-            .insert(key.clone(), payload);
-        self.image_preview
-            .payload_order
-            .retain(|queued| queued != &key);
-        self.image_preview.payload_order.push_back(key);
-        while self.image_preview.payload_order.len() > STATIC_IMAGE_INLINE_PAYLOAD_CACHE_LIMIT {
-            if let Some(stale_key) = self.image_preview.payload_order.pop_front() {
-                self.image_preview.inline_payloads.remove(&stale_key);
-            }
-        }
-    }
-
-    fn cached_prepared_static_image_for_overlay(
-        &mut self,
-        key: &StaticImageKey,
-        request: &StaticImageOverlayRequest,
-    ) -> Option<PreparedStaticImage> {
-        let dimensions = self.image_preview.dimensions.get(key).copied()?;
-        let inline_payload = if request.prepare_inline_payload {
-            Some(self.cached_static_image_inline_payload(key)?)
-        } else {
-            None
-        };
-        if self.static_image_can_use_source_path(request) {
-            return Some(PreparedStaticImage {
-                display_path: request.path.clone(),
-                dimensions,
-                inline_payload,
-            });
-        }
-
-        let display_path = self.cached_static_image_display_path(key)?;
-        Some(PreparedStaticImage {
-            display_path,
-            dimensions,
-            inline_payload,
-        })
-    }
-
-    fn remember_rendered_static_image(&mut self, key: StaticImageKey, path: PathBuf) {
-        self.image_preview.rendered_images.insert(key.clone(), path);
-        self.image_preview
-            .render_order
-            .retain(|queued| queued != &key);
-        self.image_preview.render_order.push_back(key);
-        while self.image_preview.render_order.len() > STATIC_IMAGE_RENDER_CACHE_LIMIT {
-            if let Some(stale_key) = self.image_preview.render_order.pop_front()
-                && let Some(stale_path) = self.image_preview.rendered_images.remove(&stale_key)
-            {
-                let _ = fs::remove_file(stale_path);
-            }
-        }
-    }
-
-    fn nearby_static_image_overlay_requests(
-        &self,
-        current: Option<&StaticImageOverlayRequest>,
-    ) -> Vec<StaticImageOverlayRequest> {
-        let current_path = current.as_ref().map(|request| &request.path);
-        let mut requests = self
-            .visible_entry_indices()
-            .into_iter()
-            .filter(|&index| index != self.selected)
-            .filter_map(|index| {
-                self.entries
-                    .get(index)
-                    .and_then(|entry| self.static_image_overlay_request_for_entry(entry))
-                    .map(|request| (index.abs_diff(self.selected), request))
-            })
-            .filter(|(_, request)| current_path != Some(&request.path))
-            .collect::<Vec<_>>();
-        requests.sort_by_key(|(distance, _)| *distance);
-        requests
-            .into_iter()
-            .map(|(_, request)| request)
-            .take(STATIC_IMAGE_PRELOAD_LIMIT)
-            .collect()
-    }
-
-    fn ensure_static_image_preload(
-        &mut self,
-        request: &StaticImageOverlayRequest,
-        priority: jobs::ImageJobPriority,
-    ) {
-        let key = StaticImageKey::from_request(request);
-        if self.image_preview.failed_images.contains(&key)
-            || self.image_preview.pending_prepares.contains(&key)
-            || self
-                .cached_prepared_static_image_for_overlay(&key, request)
-                .is_some()
-        {
-            return;
-        }
-
-        let job = self.image_prepare_request_for_overlay(request);
-        let submit = match priority {
-            jobs::ImageJobPriority::Current => self.scheduler.submit_image_prepare(job),
-            jobs::ImageJobPriority::Nearby => self.scheduler.submit_nearby_image_prepare(job),
-        };
-        if submit {
-            self.image_preview.pending_prepares.insert(key);
-        }
-    }
-
-    fn image_prepare_request_for_overlay(
-        &mut self,
-        request: &StaticImageOverlayRequest,
-    ) -> jobs::ImagePrepareRequest {
-        jobs::ImagePrepareRequest {
-            path: request.path.clone(),
-            size: request.size,
-            modified: request.modified,
-            target_width_px: request.target_width_px,
-            target_height_px: request.target_height_px,
-            ffmpeg_available: self.ffmpeg_available(),
-            resvg_available: self.resvg_available(),
-            magick_available: self.magick_available(),
-            force_render_to_cache: request.force_render_to_cache,
-            prepare_inline_payload: request.prepare_inline_payload,
-        }
-    }
-
-    fn active_static_image_display_target(&self) -> Option<DisplayedStaticImagePreview> {
-        let request = self
-            .active_static_image_overlay_request()
-            .or_else(|| self.active_preview_visual_overlay_request_unchecked())?;
-        let window_size = self.cached_terminal_window()?;
-        let image_dimensions = self
-            .image_preview
-            .dimensions
-            .get(&StaticImageKey::from_request(&request))
-            .copied()?;
-        Some(DisplayedStaticImagePreview::from_request(
-            &request,
-            self.static_image_display_area(&request, image_dimensions, window_size),
-            self.static_image_clear_area(&request),
-        ))
-    }
-
-    fn static_image_clear_area(&self, request: &StaticImageOverlayRequest) -> Rect {
-        if self.terminal_images.protocol == ImageProtocol::ItermInline {
-            if request.mode == StaticImageOverlayMode::Inline {
-                return self
-                    .frame_state
-                    .preview_body_area
-                    .or_else(|| self.preview_body_area())
-                    .unwrap_or(request.area);
-            }
-            return self
-                .frame_state
-                .preview_content_area
-                .unwrap_or(request.area);
-        }
-        request.area
-    }
-
-    fn preview_body_area(&self) -> Option<Rect> {
-        match (
-            self.frame_state.preview_media_area,
-            self.frame_state.preview_content_area,
-        ) {
-            (Some(media), Some(content)) => Some(union_rect(media, content)),
-            (Some(media), None) => Some(media),
-            (None, Some(content)) => Some(content),
-            (None, None) => None,
-        }
-    }
-
-    fn static_image_display_area(
-        &self,
-        request: &StaticImageOverlayRequest,
-        dimensions: RenderedImageDimensions,
-        window_size: TerminalWindowSize,
-    ) -> Rect {
-        if self.terminal_images.protocol == ImageProtocol::KittyGraphics {
-            request.area
-        } else {
-            fit_image_area(
-                request.area,
-                window_size,
-                dimensions.width_px as f32 / dimensions.height_px as f32,
-            )
-        }
-    }
-}
-
-fn union_rect(a: Rect, b: Rect) -> Rect {
-    let left = a.x.min(b.x);
-    let top = a.y.min(b.y);
-    let right = a.x.saturating_add(a.width).max(b.x.saturating_add(b.width));
-    let bottom =
-        a.y.saturating_add(a.height)
-            .max(b.y.saturating_add(b.height));
-    Rect {
-        x: left,
-        y: top,
-        width: right.saturating_sub(left),
-        height: bottom.saturating_sub(top),
     }
 }
